Agregado al context

+++ Library/PedidosDetalles.cs
@@ -11,7 +11,7 @@ public class PedidoDetalle
     public int PedidoDetalleId { get; set; }
     public int ProductoId { get; set; }
     public string ProductoNombre { get; set; }
-    public decimal PrecioUnitario { get; set; }
+    public float PrecioUnitario { get; set; }
     public int Cantidad { get; set; }
-    public decimal Total { get { return PrecioUnitario * Cantidad; } } // Calculated property
+    public float Total { get { return PrecioUnitario * Cantidad; } } // Calculated property
 }
